fastapi_factores/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
pydantic
python-dotenv
pandas
numpy
